+++ src/packs/yaml/queries/tags.scm
@@ -1,0 +1,6 @@
+; Block-style (`key: value`) and flow-style (`{key: value}`) mapping keys.
+(block_mapping_pair
+  key: (flow_node) @name) @definition.key
+
+(flow_pair
+  key: (flow_node) @name) @definition.key